+++ src/Commands/Model/SharePoint/FolderArchiveStateResult.cs
@@ -1,0 +1,15 @@
+using PnP.PowerShell.Commands.Enums;
+
+namespace PnP.PowerShell.Commands.Model.SharePoint
+{
+	public class FolderArchiveStateResult
+	{
+		public string FolderName;
+
+		public string ServerRelativeUrl;
+
+		public FolderArchiveState RequestedState;
+
+		public string MonitorUrl;
+	}
+}